commit 10.1
reworked knives to sword objects until more de-obfuscation can be done
and the knife class can be properly updated to 1.14. Some progress was
made, but a few of the methods are not yet perfected

still having issues with Forge recognizing the mod.

Next step will be to identify those and make sure that everything thus
far is  working properly

going back to school tmrw so rip free time


F